fix devise

+++ db/migrate/20131225232042_add_alchemy_roles_to_alchemy_users.rb
@@ -1,22 +1,20 @@
+# frozen_string_literal: true
+
 class AddAlchemyRolesToAlchemyUsers < ActiveRecord::Migration[4.2]
   def up
     # Updating old :roles column (since Alchemy CMS v2.6)
     if column_exists?(:alchemy_users, :roles)
       rename_column :alchemy_users, :roles, :alchemy_roles
-      change_column :alchemy_users, :alchemy_roles, :string, default: "member"
+      change_column :alchemy_users, :alchemy_roles, :string, default: 'member'
     end
 
     # Creating :alchemy_roles column for new apps.
     unless column_exists?(:alchemy_users, :alchemy_roles)
-      add_column :alchemy_users, :alchemy_roles, :string, default: "member"
+      add_column :alchemy_users, :alchemy_roles, :string, default: 'member'
     end
 
     # Renaming the index
-    if index_exists?(:alchemy_users, :roles)
-      remove_index :alchemy_users, :roles
-    end
-    unless index_exists?(:alchemy_users, :alchemy_roles)
-      add_index :alchemy_users, :alchemy_roles
-    end
+    remove_index :alchemy_users, :roles if index_exists?(:alchemy_users, :roles)
+    add_index :alchemy_users, :alchemy_roles unless index_exists?(:alchemy_users, :alchemy_roles)
   end
 end
